requirements.txt
# Core ML / data
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.7.0

# Russian NLP
pymorphy2>=0.9.1
pymorphy2-dicts-ru>=2.4.417127.4579844
nltk>=3.8.0

# Извлечение текста из документов Word и PDF (классификация и обучение из папок)
python-docx>=1.1.0
pymupdf>=1.23.0

# Опционально: загрузка обучающих данных с Hugging Face (обращения, размеченные корпуса)
datasets>=2.14.0
